fix imports for relocated ai predictive analytics
