+++ libs/ai/ai_predictive_analytics.dart
